desktop/requirements.txt
pyserial
flask
flask-cors
inputs
pyqt5
pyqtwebengine
plyer
